pyproject.toml
[tool.poetry]
name = "yolo-nas-coco"
version = "0.1.0"
description = ""
authors = ["ranhomri <ran.homri@tensorleap.ai>"]
readme = "README.md"
packages = [{include = "yolo_nas_coco"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
code-loader = "^1.0.11"
tensorflow = "2.11.0"
pillow = "^10.1.0"
pycocotools = "^2.0.6"
pyyaml = "^6.0.1"
google-cloud-storage = "^2.13.0"
code-loader-helpers = "^1.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
